refctr: step builders in their own files
