+++ core/src/lib/index.ts
@@ -4,8 +4,9 @@ export * from "./workflow-context";
 export * from "./workflow-step";
 export * from "./workflow-engine.service";
 export * from "./workflow";
+export * from "./workflow-change";
 export * from "./workflow-step-activators";
-export * from "./workflow-step-builders";
 export * from "./workflow-builder";
 
+export * from "./step-builders";
 export * from "./common-workflows.module";
